add defaults for most DB values. added logic for create payments.

+++ db/migrate/20210201060747_create_job_users.rb
@@ -2,8 +2,8 @@ class CreateJobUsers < ActiveRecord::Migration[6.1]
   def change
     create_table :job_users, id: false do |t|
 
-      t.bigint :job_id
-      t.bigint :user_id
+      t.integer :job_id
+      t.integer :user_id
 
       t.timestamps
     end
