+++ db/migrate/20210201192827_create_user_relationships.rb
@@ -2,8 +2,8 @@ class CreateUserRelationships < ActiveRecord::Migration[6.1]
   def change
     create_table :user_relationships do |t|
 
-      t.bigint :client_id
-      t.bigint :admin_id
+      t.integer :client_id
+      t.integer :admin_id
 
       t.timestamps
     end
