+++ db/migrate/20210206054950_add_contractors_to_user_relationships.rb
@@ -1,5 +1,5 @@
 class AddContractorsToUserRelationships < ActiveRecord::Migration[6.1]
   change_table(:user_relationships) do |t|
-      t.bigint :contractor_id
+      t.integer :contractor_id
   end
 end
